refactor(rag): extract getImpactContext to reduce complexity
