+++ internal/llm/rag.go
@@ -1566,99 +1566,6 @@ func isLogicFile(path string) bool {
 	return isCodeExtension(ext)
 }
 
-func (r *ragService) getImpactContext(ctx context.Context, store storage.ScopedVectorStore, repoPath string, files []internalgithub.ChangedFile, seen map[string]struct{}, mu *sync.RWMutex) string {
-	retriever := vectorstores.NewDependencyRetriever(store)
-	var impactBuilder strings.Builder
-
-	// Suggestion [internal/llm/rag.go:1447] - Parallelize dependency retrieval
-	// Dependency impact analysis is O(k log m) per call. Sequential processing scales poorly.
-	// Parallelizing with bounded concurrency reduces wall-clock time linearly.
-
-	type depRequest struct {
-		Pkg     string
-		Imports []string
-		File    internalgithub.ChangedFile
-	}
-
-	reqs := make([]depRequest, 0, len(files))
-	for _, f := range files {
-		parser, err := r.parserRegistry.GetParserForFile(f.Filename, nil)
-		if err != nil {
-			continue
-		}
-
-		// Read actual source from disk — parsing git patch syntax would fail
-		fullPath, err := r.validateAndJoinPath(repoPath, f.Filename)
-		if err != nil {
-			continue
-		}
-
-		content, err := os.ReadFile(fullPath)
-		if err != nil {
-			continue
-		}
-
-		meta, err := parser.ExtractMetadata(string(content), f.Filename)
-		if err != nil {
-			continue
-		}
-
-		reqs = append(reqs, depRequest{
-			Pkg:     meta.PackageName,
-			Imports: meta.Imports,
-			File:    f,
-		})
-	}
-
-	// Parallelize with bounded concurrency
-	const maxConcurrentDepCalls = 10
-	sem := make(chan struct{}, maxConcurrentDepCalls)
-	depResults := make(map[string][]schema.Document)
-	var depMu sync.Mutex
-
-	var wg sync.WaitGroup
-	for _, req := range reqs {
-		wg.Add(1)
-		go func(r depRequest) {
-			defer wg.Done()
-			sem <- struct{}{}
-			defer func() { <-sem }()
-
-			network, err := retriever.GetContextNetwork(ctx, r.Pkg, r.Imports)
-			if err != nil {
-				// Log warning but continue
-				return
-			}
-
-			depMu.Lock()
-			depResults[r.File.Filename] = network.Dependents
-			depMu.Unlock()
-		}(req)
-	}
-	wg.Wait()
-
-	// Process results
-	for filename, dependents := range depResults {
-		for _, doc := range dependents {
-			source, ok := doc.Metadata["source"].(string)
-			if !ok || source == "" {
-				continue
-			}
-			mu.Lock()
-			if _, exists := seen[source]; exists {
-				mu.Unlock()
-				continue
-			}
-			seen[source] = struct{}{}
-			mu.Unlock()
-
-			_, _ = impactBuilder.WriteString(fmt.Sprintf("File: %s (potential ripple effect from %s)\n---\n%s\n\n",
-				source, filename, doc.PageContent))
-		}
-	}
-	return impactBuilder.String()
-}
-
 func (r *ragService) isArchDocument(doc schema.Document) bool {
 	ct, ok := doc.Metadata["chunk_type"].(string)
 	return ok && ct == "arch"
